# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConceptSummary": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "confidence",
          "status",
          "evidence_count"
        ],
        "title": "ConceptSummary",
        "type": "object"
      },
      "DocumentSubmission": {
        "description": "POST /documents body \u2014 one raw human explanation stream.",
        "properties": {
          "context": {
            "default": "general",
            "maxLength": 100,
            "title": "Context",
            "type": "string"
          },
          "name": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "text": {
            "maxLength": 500000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "name",
          "text"
        ],
        "title": "DocumentSubmission",
        "type": "object"
      },
      "EvidenceSubmission": {
        "description": "POST /evidence body \u2014 a package without an id (the core mints it).\n\nSize caps (M50): generous for honest contributions, hostile to\npayload abuse. Artifacts are references, never blobs \u2014 the platform\nstores files.",
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "maxItems": 32,
            "title": "Artifacts",
            "type": "array"
          },
          "claim": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Claim",
            "type": "string"
          },
          "concept": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Concept",
            "type": "string"
          },
          "contributor": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Contributor",
            "type": "string"
          },
          "environment": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Environment",
            "type": "object"
          },
          "kind": {
            "default": "experiment",
            "enum": [
              "experiment",
              "replication",
              "paper",
              "transcript",
              "debate",
              "observation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "measurements": {
            "additionalProperties": true,
            "title": "Measurements",
            "type": "object"
          },
          "outcome": {
            "enum": [
              "supported",
              "challenged",
              "inconclusive"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "related_concepts": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Related Concepts",
            "type": "array"
          },
          "replicates": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replicates"
          },
          "reproduction_steps": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Reproduction Steps",
            "type": "array"
          }
        },
        "required": [
          "claim",
          "concept",
          "contributor",
          "outcome"
        ],
        "title": "EvidenceSubmission",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ALLM",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/audit": {
      "get": {
        "description": "Who changed what, when, why \u2014 metadata only, values stay\nbehind their own endpoints.",
        "operationId": "audit_trail_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Audit Trail Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Trail"
      }
    },
    "/concepts": {
      "get": {
        "operationId": "list_concepts_concepts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConceptSummary"
                  },
                  "title": "Response List Concepts Concepts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Concepts"
      }
    },
    "/concepts/{name}": {
      "get": {
        "operationId": "get_concept_concepts__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Concept Concepts  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Concept"
      }
    },
...
    "/documents": {
      "post": {
        "operationId": "submit_documents_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/DocumentSubmission"
                },
                "title": "Submissions",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Documents Documents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Documents"
      }
    },
    "/events": {
      "get": {
        "description": "The platform's live feed: events with ``seq > since``, in order.\n\nPoll with ``since`` set to the last ``seq`` you saw; ``cursor`` in\nthe response is where to resume so nothing is missed or replayed.",
        "operationId": "event_stream_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Since",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Event Stream Events Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Stream"
      }
    },
    "/evidence": {
      "post": {
        "operationId": "submit_evidence_evidence_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Evidence Evidence Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Evidence"
      }
    },
...
    "/proposals": {
      "get": {
        "operationId": "list_proposals_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Proposals Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals"
      }
    },
    "/proposals/{proposal_id}/claim": {
      "post": {
        "operationId": "claim_proposal_proposals__proposal_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Claim Proposal Proposals  Proposal Id  Claim Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Proposal"
      }
    },
    "/proposals/{proposal_id}/resolve": {
      "post": {
        "operationId": "resolve_proposal_proposals__proposal_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Proposal Proposals  Proposal Id  Resolve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Proposal"
      }
    },
...
    "/webhooks": {
      "get": {
        "description": "Subscriptions with their status \u2014 secrets never leave.",
        "operationId": "list_webhooks_webhooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Webhooks Webhooks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Webhooks"
      },
      "post": {
        "description": "Register an endpoint for the event feed. It is *proposed* and\ndelivers nothing until approved \u2014 the secret is returned once.",
        "operationId": "register_webhook_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Webhook Webhooks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Webhook"
      }
    },
...
    "/webhooks/{subscription_id}/approve": {
      "post": {
        "description": "The named-human approval gate: only now does it start delivering.",
        "operationId": "approve_webhook_webhooks__subscription_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Webhook Webhooks  Subscription Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Webhook"
      }
    },
    "/webhooks/{subscription_id}/disable": {
      "post": {
        "operationId": "disable_webhook_webhooks__subscription_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Disable Webhook Webhooks  Subscription Id  Disable Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Webhook"
      }
    },
...
  }
}